+++ tests/functions/test_retrieve_info.py
@@ -2,12 +2,12 @@
 from time import time
 from unittest import TestCase
 
-from functions.events import events as wrapped_events
-from functions.game_info import game_info as wrapped_game_info
-from functions.join_game import join_game as wrapped_join_game
-from functions.players import players as wrapped_players
-from functions.search_games import search_games as wrapped_search_games
-from functions.search_users import search_users as wrapped_search_users
+from function_app import (events as wrapped_events,
+                          game_info as wrapped_game_info,
+                          join_game as wrapped_join_game,
+                          players as wrapped_players,
+                          search_games as wrapped_search_games,
+                          search_users as wrapped_search_users)
 from utils.dtos.client import CompletedGame, Event, User, WaitingGame
 from utils.mappers.constants import EventType
 from tests.helpers import (build_request, completed_game, create_user,
